requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
crispy-bootstrap5==0.5
Django==3.2.7
django-crispy-forms==1.12.0
django-quill-editor==0.1.22
idna==3.2
Pillow==8.3.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
